Add files via upload
Such different recordings of length 30 seconds constitute the dataset of 10 different types of music genres. All the '.wav' files haven't been included due to its sheer size and some issues with the repository. The model is trained with the help of 'file.csv' CSV files which contain the details and features of all the items of the dataset.